pyproject.toml
[project]
name = "fields"
version = "1.0.0"
description = "ambient noise discord bot"
authors = [
    {name = "matt wyatt", email = "mwyatt1000@gmail.com"},
]
dependencies = [
    "discord-py>=2.3.2",
    "pynacl>=1.5.0",
    "python-dotenv>=1.0.0",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm.dev-dependencies]
dev = [
    "black>=23.10.0",
]

[tool.pdm.scripts]
start = "python src/fields"
fmt = "black src/fields"